# Discount Vec allocation time from serialization benchmark