Before:

```console
test bincode_serialize    ... bench:          18.08 ns/iter (+/- 0.18)
test postcard_serialize   ... bench:          21.13 ns/iter (+/- 0.17)
test serde_serialize      ... bench:          16.64 ns/iter (+/- 0.15)
```

After:

```console
test bincode_serialize    ... bench:           6.23 ns/iter (+/- 0.05)
test postcard_serialize   ... bench:          13.79 ns/iter (+/- 0.11)
test serde_serialize      ... bench:           6.28 ns/iter (+/- 0.05)
```