@zakarumych 2025-03-07
In some cases use have to allocate a `Vec` because it has values  in `Iterator` but `serde` can't serialize iterator unless it's `Clone`.